+++ refena_flutter/lib/src/scope.dart
@@ -172,7 +172,7 @@ class RefenaScope extends StatefulWidget implements RefenaContainer {
 
   /// Returns the actual value of a [Provider].
   @override
-  R read<N extends BaseNotifier<T>, T, R>(Watchable<N, T, R> watchable) {
+  R read<N extends BaseNotifier<T>, T, R>(BaseWatchable<N, T, R> watchable) {
     return _container.read(watchable);
   }